cleanup

+++ srcs/2darray/ft_strlistpad.c
@@ -1,27 +1,11 @@
 #include "lib2darray.h"
 
-/* Gonna ask chatGPT about it's opinion */
-// void	ft_strlistpad(char **strlist, t_ftstrpad ft_strpad)
-// {
-// 	const size_t	longest = ft_strlistfind_longest(strlist);
-// 	size_t			i;
-// 	char			*tmp;
-
-// 	i = -1;
-// 	while (strlist[++i] != NULL)
-// 	{
-// 		tmp = ft_strpad(strlist[i], longest);
-// 		free(strlist[i]);
-// 		strlist[i] = tmp;
-// 	}
-// }
-
 char	**ft_strlistpad(char **strlist, t_ftstrpad ft_strpad, const char c)
 {
-	char *const	longest = *ft_strlistfind_longest(strlist);
-	size_t		len_longest;
-	char		**newlist;
-	size_t		i;
+	char *const		longest = *ft_strlistfind_longest(strlist);
+	size_t			len_longest;
+	char			**newlist;
+	unsigned int	i;
 
 	if (longest == NULL)
 		return (ft_strlistdup(strlist));
